requirements.txt
absl-py==1.4.0
altair==4.2.2
amqp==2.6.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
asgiref==3.4.1
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.74
bcrypt==4.0.1
beautifulsoup4==4.11.1
billiard==3.6.3.0
blinker==1.8.2
blis==0.7.9
bson==0.5.10
cachetools==5.3.0
catalogue==2.0.8
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==3.0.1
click==8.1.3
cmake==3.27.4.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.0.4
contourpy==1.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbr==9.4.0
dearpygui==1.11.1
declxml==1.1.3
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.6
Django==3.1.5
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.4.0
django-currentuser==0.5.3
django-sse==0.4.1
django-timezone-field==4.2.3
djangorestframework==3.11.0
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
dpcpp-cpp-rt==2022.2.0
drf-yasg==1.17.1
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.1.1
executing==2.0.1
exif==1.6.0
fastapi==0.112.0
filelock==3.12.3
fire==0.5.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==2.5.1
flatbuffers==23.5.26
fonttools==4.38.0
future==0.18.2
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
humanfriendly==10.0
idna==3.4
ImageAug==0.1.0.post0
imageio==2.24.0
imgaug==0.4.0
imgviz==1.7.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
intel-cmplr-lib-rt==2022.2.0
intel-cmplr-lic-rt==2022.2.0
intel-opencl-rt==2022.2.0
intel-openmp==2022.2.0
ipython==8.12.3
iso8601==1.1.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keyboard==0.13.5
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
kombu==4.6.11
labelImg==1.8.6
labelme==5.3.1
langcodes==3.3.0
lit==16.0.6
lmdb==1.4.0
lxml==4.9.1
Markdown==3.4.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2022.2.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multiprocess==0.70.14
murmurhash==1.0.9
mysql-connector-python==9.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
paramiko==3.3.1
parso==0.8.4
pascal-voc-writer==0.1.4
passlib==1.7.4
pathy==0.10.1
patsy==0.5.6
pdf2docx==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
plotly==5.23.0
plum-py==0.8.7
premailer==3.10.0
preshed==3.0.8
prompt_toolkit==3.0.47
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pyclipper==1.3.0.post4
pycparser==2.20
pycryptodome==3.16.0
pydantic==1.10.4
pydeck==0.8.0
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.9.0
pylibdmtx==0.1.10
pymodbus==2.4.0
pymongo==3.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
PyMuPDF==1.20.2
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-docx==0.8.11
python-snap7==1.3
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzbar==0.1.9
QtPy==2.4.0
rapidfuzz==2.13.7
redis==4.3.4
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.3.1
rsa==4.9
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
schedule==1.2.0
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.9.3
seaborn==0.12.2
selenium==4.11.2
semver==2.13.0
serial==0.0.97
serial-tool==0.0.1
shapely==2.0.0
sip==4.19.13
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
split-folders==0.5.1
SQLAlchemy==2.0.35
sqlparse==0.4.2
srsly==2.4.5
sse==1.2
sse-starlette==1.6.1
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.1
streamlit==1.18.1
streamlit-modal==0.1.0
swig==4.1.1.post1
sympy==1.12
tbb==2021.7.1
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thinc==8.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tldextract==5.1.1
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==0.7.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.14.3
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
typer==0.7.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==4.2
ultralytics==8.0.175
uritemplate==4.1.1
urllib3==1.26.14
uuid==1.30
uvicorn==0.30.5
validators==0.20.0
vine==1.3.0
visualdl==2.4.2
wasabi==1.1.1
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
xmltodict==0.13.0
yarg==0.1.9
zipp==3.14.0
zxing-cpp==2.1.0
